% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fdc_shape.R
\docType{data}
\name{api_key}
\alias{api_key}
\title{fdc_shape_id}
\format{
An object of class \code{character} of length 1.
}
\usage{
api_key
}
\arguments{
\item{api_key}{API Key used to query the Food Data Central (FDC) database}

\item{fdc_ids}{FDC IDs to lookup}
}
\value{
data frame with item descriptors and nutritional values
}
\description{
A wrapper to query the FDC database using FDC ID numbers to get detailed information.
}
\keyword{datasets}
