% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbPct.R
\name{carbPct}
\alias{carbPct}
\title{Percentage of carbs in food}
\usage{
carbPct(fat, carbs, protein, fiber = 0, kcal = 0)
}
\arguments{
\item{fat}{grams of fat per unit of measurement (e.g. 100 g)}

\item{carbs}{grams of carbohydrates per unit of measurement (e.g. 100 g)}

\item{protein}{grams of protein per unit of measurement (e.g. 100 g)}

\item{fiber}{grams of fiber per unit of measurement (e.g. 100 g)}

\item{kcal}{total energy per unit of measurement (e.g. 100 g)}
}
\value{
percentage of energy from carbs
}
\description{
Calculates how much of the energy content comes from
carbohydrates.
}
\examples{
carbPct(57, 11, 19, 8)
}
\author{
Waldir Leoncio
}
