% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fdc_shape.R
\name{fdc_shape_id}
\alias{fdc_shape_id}
\title{fdc_shape_id}
\usage{
fdc_shape_id(api_key, fdc_ids)
}
\arguments{
\item{api_key}{API Key used to query the Food Data Central (FDC) database}

\item{fdc_ids}{FDC IDs to lookup}
}
\value{
data frame with item descriptors and nutritional values
}
\description{
A wrapper to query the FDC database using FDC ID numbers to get detailed information.
}
