% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiberGrams.R
\name{fiberGrams}
\alias{fiberGrams}
\title{Calculate the amount of fiber in food}
\usage{
fiberGrams(kcal, fat, carbs, protein)
}
\arguments{
\item{kcal}{total energy per unit of measurement (e.g. 100 g)}

\item{fat}{grams of fat per unit of measurement (e.g. 100 g)}

\item{carbs}{grams of carbohydrate per unit of measurement (e.g. 100 g)}

\item{protein}{grams of protein per unit of measurement (e.g. 100 g)}
}
\value{
Grams of fiber per unit of measurement
}
\description{
Sometimes, nutritional labels fail to inform the amount
of fiber it contains. This function helps one estimate this given
other parameters.
}
\examples{
fiberGrams(362, 17, 11, 40)
}
\author{
Waldir Leoncio
}
