% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroDistro.R
\name{macroDistro}
\alias{macroDistro}
\title{Macro distribution}
\usage{
macroDistro(fat, carbs, protein, fiber = 0)
}
\arguments{
\item{fat}{grams of fat per unit of measurement (e.g. 100 g)}

\item{carbs}{grams of carbohydrates per unit of measurement (e.g. 100 g)}

\item{protein}{grams of protein per unit of measurement (e.g. 100 g)}

\item{fiber}{grams of fiber per unit of measurement (e.g. 100 g)}
}
\value{
vector with the energy ratio from each macronutrient
}
\description{
Calculates the percentage of energy from each macronutrient.
}
\examples{
macroDistro(12, 40, 32, 1)
macroDistro(12, 40, 32)
}
\author{
Waldir Leoncio
}
