% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pct_of_day}
\alias{pct_of_day}
\title{Table of hour of day and percentage of day}
\format{
An object of class \code{data.frame} with 25 rows and 2 columns.
}
\usage{
pct_of_day
}
\description{
Table of hour of day and percentage of day
}
\keyword{datasets}
