% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalKcal.R
\name{totalKcal}
\alias{totalKcal}
\title{Total calories}
\usage{
totalKcal(fat, carbs, protein, fiber = 0)
}
\arguments{
\item{fat}{grams of fat per unit of measurement (e.g. 100 g)}

\item{carbs}{grams of carbohydrates per unit of measurement (e.g. 100 g)}

\item{protein}{grams of protein per unit of measurement (e.g. 100 g)}

\item{fiber}{grams of fiber per unit of measurement (e.g. 100 g)}
}
\value{
Total energy content per unit of measurement
}
\description{
Calculate the total caloric content of an item given the weight
of its macronutrients
}
\examples{
totalKcal(48, 1.7, 29)
}
\author{
Waldir Leoncio
}
